Modeling/requirements.txt
torch
torchvision
pandas
numpy
matplotlib
Pillow
